Make most of JS API types private
